+++ src/Rule/Rule.php
@@ -8,7 +8,10 @@ use Sabberworm\CSS\Property\Declaration;
 
 use function Safe\class_alias;
 
-/**
- * @deprecated in v9.2, will be removed in v10.0.  Use `Property\Declaration` instead, which is a direct replacement.
- */
-class_alias(Declaration::class, Rule::class);
+if (!\class_exists(Rule::class, false) && !\interface_exists(Rule::class, false)) {
+    /**
+     * @deprecated in v9.2, will be removed in v10.0.  Use `Property\Declaration` instead, which is a direct
+     *             replacement.
+     */
+    class_alias(Declaration::class, Rule::class);
+}