+++ src/RuleSet/RuleContainer.php
@@ -6,7 +6,9 @@ namespace Sabberworm\CSS\RuleSet;
 
 use function Safe\class_alias;
 
-/**
- * @deprecated in v9.2, will be removed in v10.0.  Use `DeclarationList` instead, which is a direct replacement.
- */
-class_alias(DeclarationList::class, RuleContainer::class);
+if (!\class_exists(RuleContainer::class, false) && !\interface_exists(RuleContainer::class, false)) {
+    /**
+     * @deprecated in v9.2, will be removed in v10.0.  Use `DeclarationList` instead, which is a direct replacement.
+     */
+    class_alias(DeclarationList::class, RuleContainer::class);
+}