+++ CHANGELOG.md
@@ -10,7 +10,7 @@ Please also have a look at our
 
 ### Added
 
-- Add support for modern CSS at-rules: `@layer`, `@scope`, and `@starting-style` (#????)
+- Add support for modern CSS at-rules: `@layer`, `@scope`, and `@starting-style` (#1549)
 
 ### Changed